Create logic for securely storing passwords in database
I have created the logic to securely store the password to the database by
hashing the password using bcrypt. Users can access the "users" endpoint,
including an email and password. The database itself will take care of
hashing.

In addition, users can "login" using the login endpoint, and the provided
username and password will be matched against those stored in the
database.