Add handler to update user
I added a handler to update the user when they are authenticated with the
correct JWT token.

In addition, I refactored the logic for how users are stored. Before, they
were stored using their emails as keys, but this proved to be needlessly
complicated to update upon request. The database would have to delete the
entire key/value pair and add a new one using the updated email. I have
changed the key to the user ID, which allows replacing the User struct
value instead. This is simpler because the ID remains the same. Also, due
to its consistency, it serves as a better key within the database.