Update handler logic to account for access and refresh tokens
I have updated the postLoginHandler to create access and refresh tokens on
successful login and updateUserCredsHandler to check the token in the
header and reject the request if it's a refresh token.

In addition, the server can no longer issue tokens with custom expiries.
Access tokens will expire within the hour by default, while refresh tokens
expire in 60 days.